Move the `InjextDatabaseStateIntoException` trait used in tests from `tests/` to `tests/Traits/`
